refactor: modernizar y corregir estructura backend
- Cambiar carpeta routes/ por endpoints/ para mejor organización
- Añadir Poetry como alternativa y pyproject.toml con dependencias modernas
- Corregir imports y estructura de paquetes para compatibilidad
- Actualizar configuración de base de datos SQLModel
- Añadir scripts de Poetry para fastapi dev/run
- Actualizar requirements.txt con versiones compatibles
- Añadir documentación y typing mejorado en módulos
- Corregir problemas de ejecución del servidor
- Actualizar Dockerfiles con los nuevos cambios

BREAKING CHANGE: Actualización a Python 3.13 y nueva estructura de imports
