+++ .gitignore
@@ -8,47 +8,43 @@
 *.tmp
 *.temp
 
-# === FRONTEND (Next.js / Node.js) ===
+# ===== FRONTEND (Next.js) =====
+# Dependencies
+frontend/node_modules/
+frontend/.pnp
+frontend/.pnp.js
+frontend/.yarn/install-state.gz
 
-# dependencies
-/node_modules
-node_modules/
+# Next.js
+frontend/.next/
+frontend/out/
+frontend/build/
+frontend/dist/
 
-# next.js
-/.next/
-/out/
+# Environment variables
+frontend/.env
+frontend/.env.local
+frontend/.env.development.local
+frontend/.env.test.local
+frontend/.env.production.local
 
-# production
-/build
-build/
-dist/
+# Logs
+frontend/npm-debug.log*
+frontend/yarn-debug.log*
+frontend/yarn-error.log*
 
-# debug
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-.pnpm-debug.log*
+# Runtime data
+frontend/pids
+frontend/*.pid
+frontend/*.seed
+frontend/*.pid.lock
 
-# env files
-.env*
-.env
-.env.local
-.env.production
-.env.development
+# Vercel
+frontend/.vercel
 
-# vercel
-.vercel
-
-# typescript
-*.tsbuildinfo
-next-env.d.ts
-
-# ESLint cache
-.eslintcache
-
-# Coverage
-coverage/
-*.lcov
+# TypeScript
+frontend/*.tsbuildinfo
+frontend/next-env.d.ts
 
 # === BACKEND (Python / FastAPI) ===
 
@@ -82,6 +78,10 @@ MANIFEST
 *.manifest
 *.spec
 
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
 # Unit test / coverage reports
 htmlcov/
 .tox/
@@ -108,6 +108,26 @@ venv.bak/
 backend/venv/
 backend/.venv
 
+# Poetry
+poetry.lock
+
+# Environment variables
+.env
+.env.local
+.env.development
+.env.test
+.env.production
+
+# Database
+*.db
+*.sqlite
+*.sqlite3
+mvc_proyecto.db
+backend/mvc_proyecto.db
+
+# Alembic
+alembic/versions/__pycache__/
+
 # mypy
 .mypy_cache/
 .dmypy.json
@@ -119,24 +139,32 @@ dmypy.json
 # pytype static type analyzer
 .pytype/
 
-# === BASE DE DATOS ===
-*.db
-*.sqlite
-*.sqlite3
-*.db-journal
-mvc_proyecto.db
-backend/mvc_proyecto.db
-
-# PostgreSQL
-postgresql/
-
 # === DOCKER ===
-docker-compose.override.yml
+# Docker volumes
+docker-data/
+postgres-data/
 
-# === ARCHIVOS TEMPORALES ===
-tmp/
-temp/
-.tmp/
+# === EDITORES ===
+# VSCode
+.vscode/
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+
+# IntelliJ / PyCharm
+.idea/
+*.swp
+*.swo
+*~
+
+# Sublime Text
+*.sublime-project
+*.sublime-workspace
+
+# Vim
+*.swp
+*.swo
 
 # === SISTEMA OPERATIVO ===
 # macOS
@@ -162,22 +190,42 @@ $RECYCLE.BIN/
 # Linux
 *~
 
-# === EDITORES ===
-# Vim
-*.swp
-*.swo
+# === ARCHIVOS TEMPORALES ===
+tmp/
+temp/
+.tmp/
 
-# Sublime Text
-*.sublime-workspace
-*.sublime-project
+# === LOGS ===
+logs/
+*.log
 
-# === CERTIFICADOS Y CLAVES ===
-*.pem
-*.key
-*.crt
-*.csr
+# === DOCUMENTATION ===
+docs/_build/
 
-# === ARCHIVOS ESPECÍFICOS DEL PROYECTO ===
-uploads/
-media/
-static_files/
+# === OTROS ===
+# Jupyter Notebook
+.ipynb_checkpoints
+
+# IPython
+profile_default/
+ipython_config.py
+
+# pyenv
+.python-version
+
+# Celery
+celerybeat-schedule
+celerybeat.pid
+
+# SageMath parsed files
+*.sage.py
+
+# Rope project settings
+.ropeproject
+
+# mkdocs documentation
+/site
+
+# Spyder project settings
+.spyderproject
+.spyproject
