+++ backend/.env.example
@@ -1,0 +1,22 @@
+# Variables de Base de Datos
+DATABASE_URL=postgresql://usuario:password@localhost:5432/mvc_proyecto_db
+
+# Variables de Seguridad
+SECRET_KEY=tu_clave_secreta_super_segura_cambiala_en_produccion
+ALGORITHM=HS256
+ACCESS_TOKEN_EXPIRE_MINUTES=30
+
+# Variables de la Aplicación
+API_V1_STR=/api/v1
+PROJECT_NAME=MVC Proyecto API
+DEBUG=True
+
+# Variables de CORS
+ALLOWED_ORIGINS=http://localhost:3000,http://127.0.0.1:3000
+
+# Variables de Email (opcional)
+SMTP_TLS=True
+SMTP_PORT=587
+SMTP_HOST=smtp.gmail.com
+SMTP_USER=tu_email@gmail.com
+SMTP_PASSWORD=tu_password_de_app
