+++ backend/app/__init__.py
@@ -1,0 +1,8 @@
+"""
+Este módulo inicializa el paquete de la aplicación.
+Configura la configuración global e importa los componentes necesarios.
+"""
+from .config import Settings
+
+# Crear instancia global de configuración
+settings = Settings()
