+++ backend/app/endpoints/health.py
@@ -1,5 +1,6 @@
-from fastapi import APIRouter
 from datetime import datetime
+
+from fastapi import APIRouter
 from app.schemas.schemas import HealthResponse
 
 router = APIRouter(tags=["Health"])
