+++ backend/app/main.py
@@ -1,11 +1,11 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from app.config import settings
-from app.database.database import create_db_and_tables
-from app.routes import auth_router, usuarios_router, cotizaciones_router, health_router
+from app import settings
+from app.database.database import init_db
+from app.endpoints import auth_router, usuarios_router, cotizaciones_router, health_router
 
-# Crear las tablas en la base de datos
-create_db_and_tables()
+# Inicializar la base de datos
+init_db()
 
 # Crear la aplicación FastAPI
 app = FastAPI(
@@ -29,6 +29,7 @@ app.add_middleware(
 app.include_router(health_router)
 app.include_router(auth_router, prefix=settings.API_V1_STR)
 app.include_router(usuarios_router, prefix=settings.API_V1_STR)
+app.include_router(cotizaciones_router, prefix=settings.API_V1_STR)
 
 if __name__ == "__main__":
     import uvicorn
@@ -37,5 +38,4 @@ if __name__ == "__main__":
         host="0.0.0.0",
         port=8000,
         reload=settings.DEBUG
-    )
     )
