+++ backend/app/models/__init__.py
@@ -1,1 +1,5 @@
 from .models import Usuario
+
+__all__ = [
+    "Usuario",
+]
