+++ backend/app/schemas/__init__.py
@@ -2,3 +2,14 @@ from .schemas import (
     UsuarioBase, UsuarioCreate, UsuarioResponse,
     Token, TokenData, LoginRequest, MessageResponse, HealthResponse
 )
+
+__all__ = [
+    "UsuarioBase",
+    "UsuarioCreate",
+    "UsuarioResponse",
+    "Token",
+    "TokenData",
+    "LoginRequest",
+    "MessageResponse",
+    "HealthResponse"
+]
