+++ backend/pyproject.toml
@@ -1,0 +1,65 @@
+[project]
+name = "mvc-proyecto-backend"
+version = "1.0.0"
+description = "Backend API del proyecto MVC - Sistemas de Información"
+authors = [
+    {name = "Naicriz",email = "isalazarjara@gmail.com"}
+]
+
+requires-python = ">=3.13,<4.0"
+
+dependencies = [
+    "fastapi[standard] (>=0.115.13,<0.116.0)",
+    "sqlmodel (>=0.0.24,<0.0.25)",
+    "asyncpg (>=0.30.0,<0.31.0)",
+    "redis[hiredis] (>=6.2.0,<7.0.0)",
+    "pydantic-settings (>=2.9.1,<3.0.0)",
+    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
+    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
+    "psycopg2-binary (>=2.9.10,<3.0.0)",
+    "greenlet (>=3.2.3,<4.0.0)"
+]
+
+[tool.poetry]
+packages = [
+    {include = "app"}
+]
+
+[tool.poetry.scripts]
+dev = "fastapi dev app/main.py"
+start = "fastapi run app/main.py"
+
+[tool.poetry.group.dev.dependencies]
+isort = "^6.0.1"
+flake8 = "^7.2.0"
+black = "^25.1.0"
+
+[tool.black]
+line-length = 88
+target-version = ['py313']
+
+[tool.isort]
+profile = "black"
+multi_line_output = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
+ensure_newline_before_comments = true
+line_length = 88
+
+[tool.flake8]
+max-line-length = 88
+extend-ignore = ["E203", "W503"]
+exclude = [
+    ".git",
+    "__pycache__",
+    ".venv",
+    "env",
+    "venv",
+    "dist",
+    "build"
+]
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
