+++ backend/requirements.txt
@@ -1,15 +1,14 @@
-fastapi==0.104.1
-uvicorn==0.24.0
-sqlmodel==0.0.24
-psycopg2-binary==2.9.9
-python-jose[cryptography]==3.3.0
-passlib[bcrypt]==1.7.4
-python-multipart==0.0.18
-python-dotenv==1.0.0
-pydantic==2.5.0
-pydantic-settings==2.1.0
-alembic==1.13.1
-httpx==0.25.2
-pytest==7.4.3
-pytest-asyncio==0.21.1
-pydantic[email]==2.5.0
+fastapi[standard]>=0.115.13,<0.116.0
+sqlmodel>=0.0.24,<0.0.25
+asyncpg>=0.30.0,<0.31.0
+redis[hiredis]>=6.2.0,<7.0.0
+pydantic-settings>=2.9.1,<3.0.0
+passlib[bcrypt]>=1.7.4,<2.0.0
+python-jose[cryptography]>=3.5.0,<4.0.0
+psycopg2-binary>=2.9.10,<3.0.0
+greenlet>=3.2.3,<4.0.0
+
+# Development dependencies (optional, for development environment)
+isort>=6.0.1
+flake8>=7.2.0
+black>=25.1.0
